pyproject.toml
[tool.poetry]
name = "computations"
version = "0.0.1a"
description = "Simple deffered computations service"
authors = ["Anton Siluev <burning2007@ya.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
falcon = "~2.0"
jsonschema = { version = "~3.2", extras = ["format"] }
"backports.cached-property" = { version = "~1.0", python = ">=3.6, <3.8" }
sqlalchemy = { git = "ssh://git@github.com/sqlalchemy/sqlalchemy.git", rev = "master" }
psycopg2 = "~2.8"
celery = "~4.4"
alembic = "~1.4"

[tool.poetry.dev-dependencies]
black = "20.8b1"
isort = "~5.5"
pyflakes = "~2.2"
pytest = "~6.0"
werkzeug = "~1.0"
pytest-mock = "~3.3"

[tool.black]
skip-string-normalization = true
line-length = 99

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 99

[tool.pytest.ini_options]
addopts = "--import-mode importlib"
markers = "integration: slow tests involving whole real systems interacting with each other"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
